pyproject.toml
[project]
name = "py-chatterbox-tts"
version = "0.1.0"
requires-python = "~3.11"
dependencies = [
    "numpy (>=2.3.5,<3.0.0)",
    "fastapi (>=0.124.0,<0.125.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "torchaudio (>=2.9.1,<3.0.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "torchcodec (>=0.9.0,<0.10.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "tts-webui-chatterbox-tts @ git+https://github.com/rsxdalv/chatterbox.git@1a2e63a15632a3c470c3337f44d85f923693a6b0",
    "peft (>=0.18.0,<0.19.0)",
    "setuptools (>=80.9.0,<81.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = ["taskipy (>=1.14.1,<2.0.0)"]

[tool.taskipy.tasks]
dev = "uvicorn main:app --reload --host 0.0.0.0 --port 8000"
build = "docker build -t py-chatterbox-tts ."
run = "docker run --rm -d --gpus all -p 8000:8000 --name py-chatterbox-tts py-chatterbox-tts"
